add exception to subject

+++ src/main/java/com/vstu/service/impl/SubjectService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.vstu.entity.Subject;
+import com.vstu.exceptions.AlreadyExistException;
+import com.vstu.exceptions.EntityNotFoundException;
 import com.vstu.repository.SubjectRepository;
 import com.vstu.service.interfaces.ISubjectService;
 
@@ -31,24 +33,31 @@ public class SubjectService implements ISubjectService {
 	}
 
 	@Override
-	public boolean addSubject(Subject s) {
-		if (subjectRepository.existsById(s.getId())) {
-			return false;
+	public Subject addSubject(Subject s) {
+		if (subjectRepository.existsByName(s.getName())) {
+			throw new AlreadyExistException("Subject with name: " + s.getName() + " already exists!");
 		} else {
 			subjectRepository.save(s);
-			return true;
+
 		}
+
+		return s;
 	}
 
 	@Override
 	public void updateSubject(Subject s) {
-		subjectRepository.save(s);
+		if (subjectRepository.existsById(s.getId()))
+			subjectRepository.save(s);
+		else
+			throw new EntityNotFoundException("Subject with Id:" + s.getId() + " wasn't found!");
 	}
 
 	@Override
 	public void deleteSubject(Long id) {
-		subjectRepository.deleteById(id);
-
+		if (subjectRepository.existsById(id))
+			subjectRepository.deleteById(id);
+		else
+			throw new EntityNotFoundException("Subject with Id:" + id + " wasn't found!");
 	}
 
 	@Override
